% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_bill_text.R
\name{decode_bill_text}
\alias{decode_bill_text}
\title{Decode bill text from base64 to plain text}
\usage{
decode_bill_text(bill_text)
}
\arguments{
\item{bill_text}{object returned from get_bill_text()}
}
\value{
data frame of 10 columns
}
\description{
Return a list containing bill text in a character string, and meta data
}
