% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_json_path.R
\name{find_json_path}
\alias{find_json_path}
\title{Find paths to local JSON files}
\usage{
find_json_path(base_dir, file_type)
}
\arguments{
\item{base_dir}{Character vector of the directory where the json files are stored locally}

\item{file_type}{Character string. Accepts "bill", "people", "vote", or "text"}
}
\value{
A vector of paths to .json file
}
\description{
Helper function to find local paths to json files before using the
various parsing functions in this package. Only works if directory structure
remains the same as XXX after the datasets from LegiScan,
either via API request or direct download of the bulk export.
}
