% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill.R
\name{get_bill}
\alias{get_bill}
\title{Wrapper for getBill operation}
\usage{
get_bill(api_key, bill_id)
}
\arguments{
\item{api_key}{LegiScan API key (required)}

\item{bill_id}{Identifier for bill}
}
\value{
List
}
\description{
This operation returns the primary bill detail information including sponsors,
committee references, full history, bill text and roll call information.
}
