% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_text.R
\name{get_bill_text}
\alias{get_bill_text}
\title{Wrapper for getBillText operation}
\usage{
get_bill_text(doc_id, api_key = legiscan_api_key())
}
\arguments{
\item{doc_id}{Identifier for bill text document}

\item{api_key}{Your LegiScan API key (see [legiscan_api_key()])}
}
\value{
data.frame
}
\description{
Return bill text json
}
