% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_list.R
\name{get_dataset_list}
\alias{get_dataset_list}
\title{Wrapper for getDatasetList operation}
\usage{
get_dataset_list(state = NULL, year = NULL,
  api_key = legiscan_api_key())
}
\arguments{
\item{state}{State filter (optional)}

\item{year}{Year filter (optional)}

\item{api_key}{Your LegiScan API key (see [legiscan_api_key()])}
}
\value{
data.frame
}
\description{
Returns a list of available datasets, with optional state and year filtering
}
