% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legiscan_api_key.R
\name{legiscan_api_key}
\alias{legiscan_api_key}
\title{Get or set LEGISCAN_API_KEY value}
\usage{
legiscan_api_key(set_new = FALSE)
}
\arguments{
\item{set_new}{force setting a new LegiScan API key for the current environment? Default is FALSE.}
}
\value{
atomic character vector containing the LegiScan API key
}
\description{
The API wrapper functions in this package all rely on a LegiScan API
key residing in the environment variable \code{LEGISCAN_API_KEY}. Call this
function at the beginning of a new session to set the key, or enter
`\code{file.edit("~/.Renviron")}` in the console and edit the file
`\code{LEGISCAN_API_KEY = "you_api_key"}` to set key on your local machine.
}
