% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legiscanrr-package.R
\docType{package}
\name{legiscanrr-package}
\alias{legiscanrr}
\alias{legiscanrr-package}
\title{R interface to work with LegiScan API}
\description{
Wrapper functions to get json data from LegiScan API, and cleaning etc
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fanghuiz/legiscanrr}
  \item Report bugs at \url{https://github.com/fanghuiz/legiscanrr/issues}
}

}
\author{
Author name
}
