% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legiscanrr.R
\docType{package}
\name{legiscanrr}
\alias{legiscanrr}
\alias{legiscanrr-package}
\title{R interface to work with LegiScan API}
\description{
Wrapper functions to get json data from LegiScan API, and cleaning etc
}
\author{
Author name
}
