% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill.R
\name{parse_bill}
\alias{parse_bill}
\title{Parse bill meta-data}
\usage{
parse_bill(bill)
}
\arguments{
\item{bill}{Path to bill json file or list objected returned from
\code{\link{get_bill}}}
}
\value{
A data frame of 24 columns.
For more details, see \href{../articles/parse-json.html#bill-metadata}{documentation}.
}
\description{
Parse bill meta-data from local json or API response and return a data frame.
}
\examples{
HB1 <- system.file("extdata", "bill/HB1.json", package = "legiscanrr")
HB1 <- parse_bill(HB1)
str(HB1)

}
