% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill_progress.R
\name{parse_bill_progress}
\alias{parse_bill_progress}
\title{Parse bill progress}
\usage{
parse_bill_progress(bill_json)
}
\arguments{
\item{bill_json}{Path to bill json file}
}
\value{
data.frame
}
\description{
Import bill json and parse
}
