% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill_text.R
\name{parse_bill_text}
\alias{parse_bill_text}
\title{Parse bill text json and decode from base64 to plain text}
\usage{
parse_bill_text(bill_text_json)
}
\arguments{
\item{bill_text_json}{}
}
\value{
list of 6 elements
}
\description{
Return a list containing bill text in a character string, and meta data
}
