% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_people.R
\name{parse_people}
\alias{parse_people}
\title{Parse people JSON files}
\usage{
parse_people(people_json)
}
\arguments{
\item{bill_json}{Path to people json file}
}
\value{
list
}
\description{
Import people json and parse
}
