% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rollcall_vote.R
\name{parse_rollcall_vote}
\alias{parse_rollcall_vote}
\title{Parse roll call votes}
\usage{
parse_rollcall_vote(vote_json)
}
\arguments{
\item{vote_json}{Path to vote json file}
}
\value{
A data frame of 12 columns.
For more details, see \href{../articles/parse-json.html#roll-call-votes}{documentation}.
}
\description{
Parse roll call votes for a bill from local json.
(support for parsing from API response not yet available.)
}
\examples{
vote <- system.file("extdata", "vote/154366.json", package = "legiscanrr")
vote <- parse_rollcall_vote(vote)
str(vote)

}
