% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{check_internet}
\alias{check_http_status}
\alias{check_API_response}
\title{Helper functions}
\usage{
check_internet()

check_http_status(resp)

check_API_response(content)
}
\description{
Helper functions

Check for internet

Check http status

Check API request errors
}
