% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_bill_text.R
\name{decode_bill_text}
\alias{decode_bill_text}
\title{Decode based64 encoded bill text}
\usage{
decode_bill_text(bill_text)
}
\arguments{
\item{bill_text}{List object or json file containing bill text details,
returned from \code{\link{get_bill_text}}}
}
\value{
A data frame of 10 columns, containing all the information
returned from API query, with the decoded text attached.
See \href{../articles/decode-bill-text.html}{decode bill text vignette} for more.
}
\description{
Full text of bill documents returned from LegiScan API is encoded in base64.
This function detects the document mime type, decodes the texts and turn it
as a character vector.
}
\seealso{
\code{\link{get_bill_text}}.
Text decoding utilizes R package \href{https://readtext.quanteda.io/}{readtext}.
}
