% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill.R
\name{get_bill}
\alias{get_bill}
\title{Wrapper for getBill operation}
\usage{
get_bill(bill_id, api_key = legiscan_api_key())
}
\arguments{
\item{bill_id}{Identifier for bill}

\item{api_key}{Your LegiScan API key (see [legiscan_api_key()])}
}
\value{
List
}
\description{
This operation returns the primary bill detail information including sponsors,
committee references, full history, bill text and roll call information.
}
