% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_masterlist.R
\name{get_bill_masterlist}
\alias{get_bill_masterlist}
\title{Wrapper for getMasterList operation}
\usage{
get_bill_masterlist(state = NULL, session_id = NULL,
  api_key = legiscan_api_key())
}
\arguments{
\item{state}{State abr}

\item{session_id}{Session identifier}

\item{api_key}{Your LegiScan API key (see [legiscan_api_key()])}
}
\value{
Data frame of bill information, including bill_id and bill number.
}
\description{
This operation returns a master list of bills in the given session or current state session.
}
\examples{
\dontrun{
# One state
get_bill_masterlist(state = "PA")

# Multiple states, use iterators, e.g. lapply() or map()
state_list <- c("AR", "AZ")
bills_current <- lapply(state_list, get_bill_masterlist)
}

}
