% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_masterlist.R
\name{get_bill_masterlist}
\alias{get_bill_masterlist}
\title{Wrapper for getMasterList operation}
\usage{
get_bill_masterlist(api_key, state = NULL, session_id = NULL)
}
\arguments{
\item{api_key}{LegiScan API key (required)}

\item{state}{State abr}

\item{session_id}{Session identifier}
}
\value{
Data frame of bill information, including bill_id and bill number.
}
\description{
This operation returns a master list of bills in the given session or current state session.
}
\examples{
\dontrun{
# One state
get_bill_masterlist(api_key = key, state = "AK")

# Multiple states, use iterators, e.g. \\code{[base]{lapply}} \\code{\\link[purrr]{map}}
lapply(state_list, function(x){
  get_bill_masterlist(api_key = key, state = x)
  })
}

}
