% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_text.R
\name{get_bill_text}
\alias{get_bill_text}
\title{Get bill full text}
\usage{
get_bill_text(doc_id, api_key = legiscan_api_key())
}
\arguments{
\item{doc_id}{Identifier for bill text document}

\item{api_key}{Your LegiScan API key (see \code{\link{legiscan_api_key}})}
}
\value{
A list, with \code{doc_id} as the unique identifier.
}
\description{
This is the API wrapper function for the getBillText operation.
It returns a copy of base64 encoded bill text document, along with other
information such as document date, type, MIME type etc.
}
\seealso{
\code{\link{legiscan_api_key}},
\href{https://legiscan.com/gaits/documentation/legiscan}{LegiScan API manual}.
For decoding the base64 encoded document, see \code{\link{decode_bill_text}}
and the related \href{../articles/decode-bill-text.html}{vignette}.
}
