% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_text.R
\name{get_bill_text}
\alias{get_bill_text}
\title{Wrapper for getBillText operation}
\usage{
get_bill_text(api_key, doc_id)
}
\arguments{
\item{api_key}{LegiScan API key (required)}

\item{doc_id}{Identifier for bill text document}
}
\value{
data.frame
}
\description{
Return bill text json
}
