% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_text.R
\name{get_bill_text}
\alias{get_bill_text}
\title{Wrapper for getBillText operation}
\usage{
get_bill_text(bill_json, api_key)
}
\arguments{
\item{bill_json}{}

\item{api_key}{LegiScan API key (required)}

\item{doc_id}{Bill text doc_id}
}
\value{
data.frame
}
\description{
Return bill text and save to disk
}
