% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Wrapper for getDataset operation}
\usage{
get_dataset(api_key, dataset_list_object, save_data_to)
}
\arguments{
\item{api_key}{LegiScan API key (required)}

\item{dataset_list_object}{Object returned from get_dataset_list()}

\item{save_data_to}{}
}
\value{
NULL. No return value. Dataset will be unzipped to local disk.
}
\description{
Unzip a single ZIP archive of the requested session dataset.
}
