% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Wrapper for getDataset operation}
\usage{
get_dataset(api_key, dataset_list, save_to_dir = ".")
}
\arguments{
\item{api_key}{LegiScan API key (required)}

\item{dataset_list}{Object returned from get_dataset_list()}

\item{save_to_dir}{Local directory to save the unzipped dataset.
Default to working directory "."}
}
\value{
NULL. No return value. Dataset will be unzipped to local disk.
}
\description{
Unzip a single ZIP archive of the requested session dataset.
}
