% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_list.R
\name{get_dataset_list}
\alias{get_dataset_list}
\title{Get list of available datasets}
\usage{
get_dataset_list(state = NULL, year = NULL,
  api_key = legiscan_api_key())
}
\arguments{
\item{state}{State filter (optional)}

\item{year}{Year filter (optional)}

\item{api_key}{Your LegiScan API key (see \code{\link{legiscan_api_key}})}
}
\value{
List of dataset information.
}
\description{
This is the API wrapper function for the getDatasetList operation.
It returns a list of available datasets, with optional \code{state}
and/or \code{year} filtering.
}
\details{
This opertation does not return the dataset itself. Rather, it gives the
dataset meta-data, and the \code{session_id} and \code{access_key} required
for the \code{\link{get_dataset}} function to retrieve the dataset.

\itemize{
  \item If neither \code{state} nor \code{year} is given, returns a full list of all available dataset
  \item If only \code{state} is given, returns datasets for that state from all available years
  \item If only \code{year} is given, returns datasets for that year from all 50 states,
plus DC and Congress.
}
}
\seealso{
\code{\link{get_dataset}}, \code{\link{legiscan_api_key}},
\href{https://legiscan.com/gaits/documentation/legiscan}{LegiScan API manual}
}
