% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill.R
\name{parse_bill}
\alias{parse_bill}
\title{Parse bill JSON files}
\usage{
parse_bill(bill_json)
}
\arguments{
\item{bill_json}{Path to bill json file}
}
\value{
list
}
\description{
Import bill json and parse
}
