% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill_progress.R
\name{parse_bill_progress}
\alias{parse_bill_progress}
\title{Parse bill progress events}
\usage{
parse_bill_progress(bill)
}
\arguments{
\item{bill}{Path to bill json file or list objected returned from
\code{\link{get_bill}}}
}
\value{
A data frame of 3 columns.
For more details, see \href{../articles/parse-json.html#bill-progress}{documentation}.
}
\description{
Parse bill progress data from local json or API response and return a data frame.
}
\details{
Progress event is represented in numeric id.
See \href{../articles/lookup_table.html#bill-progress-id}{lookup table} for description.
}
\examples{
HB1 <- system.file("extdata", "bill/HB1.json", package = "legiscanrr")
HB1 <- parse_bill_progress(HB1)
str(HB1)

}
