% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bill_sponsor.R
\name{parse_bill_sponsor}
\alias{parse_bill_sponsor}
\title{Parse bill sponsors}
\usage{
parse_bill_sponsor(bill_json)
}
\arguments{
\item{bill_json}{Path to bill json file}
}
\value{
data.frame
}
\description{
Import bill json and parse
}
