% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_people.R
\name{parse_people}
\alias{parse_people}
\title{Parse legislator information}
\usage{
parse_people(people_json)
}
\arguments{
\item{people_json}{Path to people json file}
}
\value{
A data frame of 20 columns.
For more details, see \href{../articles/parse-json.html#legislator-information}{documentation}.
}
\description{
Parse bill sponsor information from local json. (support for parsing from API response
not yet available.)
}
\examples{
person_8630 <- system.file("extdata", "people/8630.json", package = "legiscanrr")
person_8630 <- parse_people(person_8630)
str(person_8630)

}
