% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_person_vote.R
\name{parse_person_vote}
\alias{parse_person_vote}
\title{Parse individual legislator votinf record}
\usage{
parse_person_vote(vote_json)
}
\arguments{
\item{vote_json}{Path to vote json file}
}
\value{
data.frame
}
\description{
Import vote json and parse
}
