% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rollcall_vote.R
\name{parse_rollcall_vote}
\alias{parse_rollcall_vote}
\title{Parse roll call votes}
\usage{
parse_rollcall_vote(vote_json)
}
\arguments{
\item{bill_json}{Path to vote json file}
}
\value{
data.frame
}
\description{
Import vote json and parse
}
